requirements.txt
feedparser>=6.0.0
requests>=2.25.0
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.2
Flask-WTF>=1.1.1
email-validator>=2.0.0
python-dotenv>=1.0.0
werkzeug>=2.0.0 